requirements.txt
Flask
Flask-WTF
Flask-SQLAlchemy
Flask-Migrate
aiohttp
Faker
beautifulsoup4
WTForms
